bot-main 2/requirements.txt
aiodns==3.2.0
aiohttp==3.9.3
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
binance==0.3
bta-lib==1.0.0
ccxt==4.3.40
ccxtpro==1.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
executing==2.0.1
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
gevent==24.2.1
greenlet==3.0.3
html5lib==1.1
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
matplotlib-inline==0.1.6
mplfinance==0.12.10b0
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pandas_ta==0.3.14b0
parso==0.8.3
peewee==3.17.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
ta==0.11.0
tenacity==8.4.1
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.1
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
wcwidth==0.2.12
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
yfinance==0.2.41
zope.event==5.0
zope.interface==6.4.post2
